soc: qcom: qmi: Add snapshot of QMI driver
This snapshot is taken as of msm-4.14 'commit <101274def507> ("Merge
"ARM: dts: msm: Change glink intent size for diag on sdmmagpie"")'.

This change brings the QMI driver up to data with fixes from msm-4.14.

+++ include/linux/soc/qcom/qmi.h
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-2.0
 /*
- * Copyright (c) 2012-2014, The Linux Foundation. All rights reserved.
+ * Copyright (c) 2012-2014, 2018 The Linux Foundation. All rights reserved.
  * Copyright (c) 2017, Linaro Ltd.
  */
 #ifndef __QMI_HELPERS_H__
@@ -166,7 +166,7 @@ struct qmi_ops {
 struct qmi_txn {
 	struct qmi_handle *qmi;
 
-	int id;
+	u16 id;
 
 	struct mutex lock;
 	struct completion completion;
